Move `Contract` related classes to separate directory
