+++ starknet_py/tests/e2e/block_test.py
@@ -1,6 +1,5 @@
 import pytest
 
-from starknet_py.contract import Contract
 from starknet_py.net.account.base_account import BaseAccount
 from starknet_py.net.client_models import (
     PendingStarknetBlock,
@@ -8,6 +7,7 @@ from starknet_py.net.client_models import (
     StarknetBlock,
     StarknetBlockWithTxHashes,
 )
+from starknet_py.net.contract.contract import Contract
 from starknet_py.tests.e2e.fixtures.constants import MAX_FEE
