+++ starknet_py/tests/e2e/contract_interaction/v1_interaction_test.py
@@ -3,7 +3,7 @@ import sys
 import pytest
 
 from starknet_py.cairo.felt import decode_shortstring, encode_shortstring
-from starknet_py.contract import Contract
+from starknet_py.net.contract.contract import Contract
 from starknet_py.tests.e2e.fixtures.constants import MAX_FEE
 from starknet_py.tests.e2e.fixtures.contracts import deploy_v1_contract
