+++ starknet_py/tests/e2e/docs/code_examples/test_contract.py
@@ -1,8 +1,8 @@
 # pylint: disable=unused-variable
 import pytest
 
-from starknet_py.contract import Contract
 from starknet_py.net.account.account import Account
+from starknet_py.net.contract.contract import Contract
 from starknet_py.net.full_node_client import FullNodeClient
 from starknet_py.net.models import StarknetChainId
 from starknet_py.net.signer.stark_curve_signer import KeyPair
