+++ starknet_py/tests/e2e/docs/code_examples/test_prepared_function_call.py
@@ -1,7 +1,7 @@
 # pylint: disable=unused-variable
 import pytest
 
-from starknet_py.contract import Contract
+from starknet_py.net.contract.contract import Contract
 
 
 @pytest.mark.asyncio
