+++ starknet_py/tests/e2e/docs/guide/test_handling_client_errors.py
@@ -5,8 +5,8 @@ import pytest
 async def test_handling_client_errors(account):
     # pylint: disable=import-outside-toplevel
     # docs: start
-    from starknet_py.contract import Contract
     from starknet_py.net.client_errors import ClientError
+    from starknet_py.net.contract.contract import Contract
 
     try:
         contract_address = "1"  # Doesn't exist
